Use simple JSON API to get package releases.
This replaces the XML-RPC mechanism (which is deprecated and heavily
rate-limited, as documented at
https://warehouse.pypa.io/api-reference/xml-rpc.html) with the simple
JSON API (PEP 691, https://peps.python.org/pep-0691/).

This is a minimalistic implementation with almost no error handling, but
in this context it should behave about the same as the old XML-RPC code
(which also had almost no error handling).